+++ python/ray/tests/test_gcs_pubsub.py
@@ -2,85 +2,11 @@ import asyncio
 import re
 import sys
 import threading
-from collections import deque
 
-import grpc
 import pytest
 
 import ray
-from ray._private.gcs_pubsub import (
-    GcsAioNodeInfoSubscriber,
-    GcsAioResourceUsageSubscriber,
-)
-
-
-class _FakeRpcError(grpc.RpcError):
-    def __init__(self, status_code):
-        self._status_code = status_code
-
-    def code(self):
-        return self._status_code
-
-
-def _make_aio_subscriber(subscriber_type):
-    subscriber = subscriber_type.__new__(subscriber_type)
-    subscriber._subscriber_id = b"\x01" * 28
-    subscriber._last_batch_size = 0
-    subscriber._max_processed_sequence_id = 0
-    subscriber._publisher_id = b""
-    subscriber._queue = deque()
-    subscriber._close = asyncio.Event()
-    return subscriber
-
-
-@pytest.mark.asyncio
-async def test_aio_node_subscriber_detects_lost_subscription():
-    subscriber = _make_aio_subscriber(GcsAioNodeInfoSubscriber)
-
-    async def unavailable_poll(request, timeout=None):
-        assert request.reject_if_subscriber_missing
-        raise _FakeRpcError(grpc.StatusCode.UNAVAILABLE)
-
-    subscriber._poll_call = unavailable_poll
-
-    with pytest.raises(_FakeRpcError):
-        await subscriber.poll(batch_size=1)
-
-
-@pytest.mark.asyncio
-async def test_aio_resource_subscriber_suppresses_unavailable():
-    subscriber = _make_aio_subscriber(GcsAioResourceUsageSubscriber)
-
-    async def unavailable_poll(request, timeout=None):
-        assert not request.reject_if_subscriber_missing
-        raise _FakeRpcError(grpc.StatusCode.UNAVAILABLE)
-
-    subscriber._poll_call = unavailable_poll
-
-    await subscriber._poll()
-
-
-@pytest.mark.asyncio
-async def test_aio_poll_cancels_inflight_rpc_when_closed():
-    subscriber = _make_aio_subscriber(GcsAioResourceUsageSubscriber)
-    poll_started = asyncio.Event()
-    poll_cancelled = asyncio.Event()
-
-    async def blocking_poll(request, timeout=None):
-        poll_started.set()
-        try:
-            await asyncio.Event().wait()
-        finally:
-            poll_cancelled.set()
-
-    subscriber._poll_call = blocking_poll
-    poll_task = asyncio.create_task(subscriber._poll())
-    await poll_started.wait()
-
-    subscriber._close.set()
-    await poll_task
-
-    assert poll_cancelled.is_set()
+from ray._private.gcs_pubsub import GcsAioResourceUsageSubscriber
 
 
 def test_publish_and_subscribe_error_info(ray_start_regular):
